backend/requirements.txt
# Backend requirements
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
passlib[bcrypt]
pyjwt
requests
